Add files via upload
This chart is my guide to understanding crow's feet notation. Technically, this chart is incorrect because there is maximum and minimum cardinality which actually refers to the minimum and maximum objects that can/will be related to another object(s), but when beginning to learn crow's feet notation, this is what made sense to me when learning optionality, cardinality, and the crow's feet symbols.